feat(RelativeRangeDatePicker): added interactive tooltip and header
