+++ src/components/RelativeRangeDatePicker/components/PickerDialog/i18n/en.json
@@ -4,5 +4,6 @@
   "\"From\" can't be after \"To\".": "\"From\" can't be after \"To\".",
   "From": "From",
   "To": "To",
-  "Apply": "Apply"
+  "Apply": "Apply",
+  "Specify the interval": "Specify the interval"
 }
