+++ src/components/RelativeRangeDatePicker/components/PickerDialog/i18n/ru.json
@@ -4,5 +4,6 @@
   "\"From\" can't be after \"To\".": "Значение «От» не может быть позже чем «До».",
   "From": "От",
   "To": "Дo",
-  "Apply": "Применить"
+  "Apply": "Применить",
+  "Specify the interval": "Выберите интервал"
 }
